[#45]refactor: 디렉토리 리네이밍 및 리스트 리팩터링 - @components/guide/product -> @components/guide/template으로 리네이밍 - SessionList, TemplateList 리팩터링
